Fixed most of the warnings in dcmjpeg.
Compiled with extensive warning options except -Wtraditional and corrected
most of the problems. Tested with mingw-w64-4.8.1, gcc-4.7.2 (static and shared),
Visual Studio 2013 Express, Visual Studio 2012 64 bit and clang-w64-3.5.0.
Tested compression / decompression with various different settings for different
compilers on appropriate DICOM images ("cabinet" and "car97_cd").
Most of the remaining warnings in dcmjpeg are caused by files included from
dcmimage and will be resolved in a different commit.

+++ dcmjpeg/include/dcmtk/dcmjpeg/djcodece.h
@@ -1,6 +1,6 @@
 /*
  *
- *  Copyright (C) 2001-2011, OFFIS e.V.
+ *  Copyright (C) 2001-2014, OFFIS e.V.
  *  All rights reserved.  See COPYRIGHT file for details.
  *
  *  This software and supporting documentation were developed by
@@ -213,7 +213,7 @@ protected:
    */
   static OFCondition togglePlanarConfiguration8(
     Uint8 *pixelData,
-    const unsigned long numValues,
+    const size_t numValues,
     const Uint16 samplesPerPixel,
     const Uint16 oldPlanarConfig);
 
@@ -230,7 +230,7 @@ protected:
    */
   static OFCondition togglePlanarConfiguration16(
     Uint16 *pixelData,
-    const unsigned long numValues,
+    const size_t numValues,
     const Uint16 samplesPerPixel,
     const Uint16 oldPlanarConfig);
